#54: Changes color values / types due to redesigning of color.

+++ WinBoyEmulator.GameBoy/GPU/Palette.cs
@@ -21,16 +21,13 @@ using System.Threading.Tasks;
 namespace WinBoyEmulator.GameBoy.GPU
 {
     /// <summary>
-    /// Class Palette. <para />
-    /// Every types can be here safely byte.
-    /// Reason is, that we are saving handling here RGB values.
-    /// Size of RGB value is one byte.
+    /// Class Palette.
     /// </summary>
     internal class Palette
     {
-        public byte[] Background { get; set; }
-        public byte[] Object1 { get; set; }
-        public byte[] Object2 { get; set; }
+        public int[] Background { get; set; }
+        public int[] Object1 { get; set; }
+        public int[] Object2 { get; set; }
 
         /// <summary>
         /// Constructor.
@@ -38,9 +35,9 @@ namespace WinBoyEmulator.GameBoy.GPU
         /// <param name="colorsInPalette">Amount of colors in the palette.</param>
         public Palette(int colorsInPalette)
         {
-            Background = new byte[colorsInPalette];
-            Object1 = new byte[colorsInPalette];
-            Object2 = new byte[colorsInPalette];
+            Background = new int[colorsInPalette];
+            Object1 = new int[colorsInPalette];
+            Object2 = new int[colorsInPalette];
         }
     }
 }
